Let's pass all inputs around explicitely.
This factors out all inputs into a separate class that gets passed around instead of a global. This way dataflow is easier to see and also its always clear if something is correctly initialized or not (if you have it its initialized).  And we are not relying on module initialization order.

In the same breath I've changed generator.Generator -> generator.Result and now it's really just that "a big container of computed values".

And Generator.calculate / __post_init__ has become simple functions.

Why? This is steps towards an eventual world where all data flow is explicit.

In terms of functionality this makes it possible to actually provide different ags / year and entries to the generator without changing a source file. Which was a requirement before we use the generator in the website.