This aligns the control and the dataflow. How? Very simply. Before we allocated the big tree of result values before all computations were done and then passed that into every calculation. Each calculation was then in charge of filling out the variables it was in charge of.
Now every sector's calc functions gets passed in the sectors it needs as inputs (and only those) and then returns the sector it computed.

And only at the very end everything is assembled into one big record of records (dataclass of dataclasses in python parlance).

Note that we should really do the same exercise again for each sector... But that will be an exercise for another day.