Get us closer to type checking.
This ignores the type checker to ignore that all variables get initialized with None. This is not ideal (as it now can happen that we use an uninitialized member variable but the type checker believes all is fine), but the proper way to fix this is to only create the records when we have already computed the values. Which is a lot more work than this quick hack.

Why do I want to be able to run the type checker? Well two reasons

a) Static correctness checks means one has to be less careful when the code is changed.

b) The types are part of our public API. In particular once we have a REST service.  And static types can act as a part of the documentation that the type checker will force us to keep up to date!

On the bright side now the only remaining errors that pyright shows in basic type checking mode seem to be real mistakes.