Call pyright as part of ready-to-rock.sh
This means we now have more static checking that the code works and that type declarations and code are in sync. Which is good because we have such a big type as part of our public API.

Also fixes two more places where pyright didn't pass.  One of them the usual let's pretend it's okay that we initialize variables with None but claim they are floats.

The other is a place where set pct changed variable to None instead of 0, when the term we divide by is 0. This seems to be a leftover of the time when we hadn't really decided that those variables can just be 0.  So this code changes that accordingly. Hence the updated expectations.