Handle a few more of the possible tree diffs. And share more of the UI code.
This simplifies the diff logic. That simplification isn't exactly free as the new type can't actually "properly" represent some theoretically possible diffs (one element on one side and tree at the same side on the other). But it's good enough for what we do and means writing the visualisation code for the diffs becomes easier (so now THAT part of the code handles more cases than we previously did).