Revert "Temporary patch to check that with tracing enabled the regression tests still produce the same numbers."
This reverts commit 8eced4092539ba0d2c63e8fd30a4bf617b77b4ad.

Reverted because one cannot monkeypatch in the same executable we run the testing for not monkeypatch'ed code and expect the tests to not be influenced.

So for now this was a one of to convince myself that the traced numbers are identical to untraced numbers until I figured out a good plan to do this all the time.