Bugfixes in transport calculations. Found while preparing/testing the data update for 2021 by exchanging population file only and looking at the differences.
a) We used the wrong fact (foot instead of cycl modal split) for 2018 other_cylc_transport_capacity_pkm.  Didn't matter in the end because both are 1.0, but still potential for future bug if one of the values changed.

b) when calculating t30.other_foot.change_transport_capacity_pkm we computed the change against cycl(!) instead of foot.  This meant this value was just wrong.